% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparision.R
\name{Comparision}
\alias{Comparision}
\title{use of ggplot}
\usage{
Comparision (x) 
for (i in 1:20) { 
simulated_data[[i]] <- trees[[i]]$root.edge 
}
}
\arguments{
\item{simulated_data}{the simulated data}
}
\value{
plot
}
\description{
use of ggplot
}
