% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{multi}
\alias{multi}
\title{Multiply together two numbers}
\usage{
multi(4, 5)
}
\arguments{
\item{x}{A number}

\item{y}{Another number}
}
\value{
product The product of multiplying x and y
}
\description{
Multiply together two numbers
}
