% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior trace.R
\name{posterior_trace}
\alias{posterior_trace}
\title{Data Visualization
use function rangeplot.asymmetric to simulate tree for data visualization.}
\usage{
i = 1
for (i in 1:20) {
posterior_trace(i)
}
}
\arguments{
\item{n}{number from 1 to 20}
}
\value{
trees
}
\description{
Data Visualization
use function rangeplot.asymmetric to simulate tree for data visualization.
}
