% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior trace.R
\name{treeplot}
\alias{treeplot}
\title{Data Visualization
use function rangeplot.asymmetric to simulate tree for data visualization.}
\usage{
treeplot(1)
}
\arguments{
\item{n}{number from 1 to 20}
}
\value{
trees
}
\description{
Data Visualization
use function rangeplot.asymmetric to simulate tree for data visualization.
}
