% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{Simulation}
\alias{Simulation}
\title{Data simulations
use sim.fbd.taxa to simulate data.}
\usage{
Simulation(n = 10, numbsim = 2, lambda = 0.01806175, mu = 1.078013e-03, psi = 5.172419e-02, frac = .01))
}
\arguments{
\item{n}{Number of extant sampled tips.}

\item{numbsim}{Number of trees to simulate.}

\item{frac}{fraction.}
}
\value{
result
}
\description{
Data simulations
use sim.fbd.taxa to simulate data.
}
