% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{Simulation}
\alias{Simulation}
\title{Data simulations
use sim.fbd.taxa to simulate data.}
\usage{
Simulation (10, 2 , 0.01)
}
\arguments{
\item{n}{Number of extant sampled tips.}

\item{numbsim}{Number of trees to simulate.}

\item{frac}{fraction.}
}
\value{
trees\link{i}
}
\description{
Data simulations
use sim.fbd.taxa to simulate data.
}
